GPT-4
I switched the model to GPT-4 since the OpenAI Gods seem to have granted
my request to join the beta program, for reasons I don't understand
since I think it's all a steaming pile of shit.

I upped the max_tokens to 600 and also added to the prompt to prevent it
from hallucinating a date which doesn't match the current date.